vim: set nofsync
There's no point in forcing a flush when the OS can't support it.
It affects swapfiles which may prevent large file writes.
It happened to me during work when using Vim on SSH.
I couldn't save because of swapfile so I had to start over.
fsync() guaranteed that I can never save.