vim: win: drop truecolor support
ConEmu is the only Windows-only non-Election terminal
that is worth using on Windows.
It has truecolor support but Vim doesn't support it officialy.
ConEmu truecolor can break between patch releases
and it's a PITA to hack Vim to work in ConEmu..
Neovim can detect it correctly thanks to libuv.

Happy to see the day that nvim's TUI works better than vim on Windows.